backend/requirements.txt
acres==0.5.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.7
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.6
distlib==0.4.0
etelemetry==0.3.1
fastapi==0.135.3
filelock==3.20.0
greenlet==3.3.2
h11==0.16.0
httplib2==0.31.2
identify==2.6.15
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
looseversion==1.3.0
lxml==6.0.3
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
nodeenv==1.9.1
numpy==2.4.4
packaging==25.0
pandas==3.0.2
pathlib==1.0.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
prov==2.1.1
puremagic==2.2.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.12.1
PyMuPDF==1.27.2.2
PyMySQL==1.1.2
pyparsing==3.3.2
pypdf==6.9.2
PyPDF2==3.0.1
pytest==9.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
requests==2.33.1
scipy==1.17.1
simplejson==3.20.2
six==1.17.0
SQLAlchemy==2.0.48
sqlalchemy-orm==1.2.10
starlette==1.0.0
tools==1.0.26
traits==7.1.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==20.35.4
